\name{tomatoGeneChipORGANISM}
\alias{tomatoGeneChipORGANISM}
\title{The Organism for tomatoGeneChip}
\description{
  tomatoGeneChipORGANISM is an R object that contains a single item: a
  character string that names the organism for which tomatoGeneChip was built.
}
\details{
  Although the package name is suggestive of the organism for which it
  was built, tomatoGeneChipORGANISM provides a simple way to programmatically
  extract the organism name.
}
\examples{
  tomatoGeneChipORGANISM
}
\keyword{datasets}

